creates vehicle controller

+++ backend/src/main/java/com/cdac/acts/logistics_v1/service/impl/UserServiceImpl.java
@@ -86,9 +86,14 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUser(Long id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return true; 
+        }
+        return false; 
     }
+
 
     @Override
     public List<UserResponseDTO> findUsersByRole(String role) {
